# laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "JBI-3",
        "contact": {
            "email": "admin@mail.ru"
        },
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8083/api/",
            "description": "JBI-3"
        }
    ],
    "paths": {
...
        "/api/v1/admin/news/": {
            "get": {
                "tags": [
                    "Admin/News"
                ],
                "summary": "Получение списка новостей",
                "description": "Возвращает список новостей",
                "operationId": "6130a74ee65a41114154cad979bbcab6",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NewsResource"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Admin/News"
                ],
                "summary": "Добавление новой новости",
                "operationId": "277b734c51332ad4e1a5130ee7b0b249",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateNewsRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NewsResourceOne"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/admin/news/{id}": {
            "get": {
                "tags": [
                    "Admin/News"
                ],
                "summary": "Получение новости",
                "operationId": "904baa3704e71fc5901dd6587ee9233d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id новости",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NewsResourceOne"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/admin/news/${id}": {
            "put": {
                "tags": [
                    "Admin/News"
                ],
                "summary": "Обновление данных новости",
                "operationId": "550f0b9476fe3b98e94306538af45fda",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id новости",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateNewsRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NewsResourceOne"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Admin/News"
                ],
                "summary": "Удаление новости",
                "operationId": "4150275cf8425f436965e0b237fe50e8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id новости",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NewsResourceOne"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "News": {
                "title": "News",
                "description": "Модель новости",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "id",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "title": {
                        "title": "title",
                        "description": "Название новости",
                        "type": "string",
                        "example": "JBI"
                    },
                    "type": {
                        "title": "type",
                        "description": "Тип новости",
                        "type": "string",
                        "example": "Тип новости"
                    },
                    "category_id": {
                        "title": "category_id",
                        "description": "category_id",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "is_active": {
                        "title": "is_active",
                        "description": "Активность новости",
                        "type": "boolean",
                        "example": true
                    },
                    "slug": {
                        "title": "slug",
                        "description": "Преобразованное название для ссылки",
                        "type": "string",
                        "example": "test-news"
                    },
                    "date": {
                        "title": "date",
                        "description": "Дата новости",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-09-12"
                    },
                    "preview_img": {
                        "title": "preview_img",
                        "description": "Превью картинка",
                        "type": "string",
                        "example": "[]"
                    },
                    "detail_img": {
                        "title": "detail_img",
                        "description": "Детальная картинка",
                        "type": "string",
                        "example": "[]"
                    },
                    "content": {
                        "title": "content",
                        "description": "Контент",
                        "type": "string",
                        "example": "[]"
                    },
                    "preview_text": {
                        "title": "preview_text",
                        "description": "Превью текст",
                        "type": "string"
                    },
                    "detail_text": {
                        "title": "detail_text",
                        "description": "Детальный текст",
                        "type": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "News"
                }
            },
...
            "CreateNewsRequest": {
                "title": "CreateNewsRequest",
                "description": "Тело запроса",
                "required": [
                    "title",
                    "is_active"
                ],
                "properties": {
                    "title": {
                        "title": "title",
                        "description": "Название новости",
                        "type": "string",
                        "example": "JBI-3"
                    },
                    "is_active": {
                        "title": "is_active",
                        "description": "Активность новости",
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
            "UpdateNewsRequest": {
                "title": "UpdateNewsRequest",
                "description": "Тело запроса",
                "required": [
                    "title",
                    "is_active"
                ],
                "properties": {
                    "title": {
                        "title": "title",
                        "description": "Название новости",
                        "type": "string",
                        "example": "JBI-3"
                    },
                    "type": {
                        "title": "type",
                        "description": "Тип новости",
                        "type": "string",
                        "example": "маленькая|большая"
                    },
                    "is_active": {
                        "title": "is_active",
                        "description": "Активность новости",
                        "type": "boolean",
                        "example": true
                    },
                    "date": {
                        "title": "date",
                        "description": "Дата новости",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-09-12"
                    },
                    "preview_img": {
                        "title": "preview_img",
                        "description": "Превью картинка",
                        "type": "string",
                        "example": "[]"
                    },
                    "detail_img": {
                        "title": "detail_img",
                        "description": "Детальная картинка",
                        "type": "string",
                        "example": "[]"
                    },
                    "content": {
                        "title": "content",
                        "description": "Контент",
                        "type": "string",
                        "example": "[]"
                    },
                    "preview_text": {
                        "title": "preview_text",
                        "description": "Превью текст",
                        "type": "string",
                        "example": "[]"
                    },
                    "detail_text": {
                        "title": "detail_text",
                        "description": "Детальный текст",
                        "type": "string",
                        "example": "[]"
                    }
                },
                "type": "object"
            },
...
            "NewsResource": {
                "title": "NewsResource",
                "description": "News resource",
                "properties": {
                    "none_data": {
                        "title": "Data",
                        "description": "data",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/News"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "NewsResource"
                }
            },
...
            "NewsResourceOne": {
                "title": "NewsResourceOne",
                "description": "News resource",
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/News"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "NewsResourceOne"
                }
            },
...
        }
    }
}